walkman-=firmata controlled arduino;+=custom c programm
This large commit drops the usage of firmata in python in order to
control electromagnets.
Instead of this i use a custom C program now, which can communicate
with python over pyserial and a custom protocol.

The reason for this step is very simple: python ate too much cpu
(because we had to call functions on audio time, which is obviously
a very bad idea). We couldn't control 9 electromagnets with this
old approach. The new approach is a lot more code, but also works
much better.