Replace astral moon phase calculation by custom calc
Astral gives us how long the moon takes around the earth (~27days), but
we need to know how long *it looks like for human beings* (~29days). So
we replaced this. We also added a custom moon index calculator, which
ensures that each day has exactly one specific moon phase index: this is
important for mapping moon phase index to scores.