render/notation: Explicitly parse tag_tuple for better control
The tag_tuple is used to extract all event placements with a given
tag. This should always only be the instrument of the part book and the
pclock, because otherwise - if a tag_to_event_placement_converter
provides more tags - which may be needed if we add entries which
encompass multiple instruments - the score rendered would print all
entries, even those of instruments which currenlty don't matter for the
given event placement.