f0/arduino: Drop DSP AUTO envelope, split tones (fix duration)
If we have very long notes the duration wasn't represented correctly if
it was very long (because we only use C "int" for duration).
This could be fixed by switching to "long", but then the duration is
also clipped at a later point (so protocol is still unreliable) + we
waste more disk space on the arduino. So it's better to simply split all
notes when converting them to f0 procotol. In this we can also easily
apply micro changes.