etc/arduino/f0: Add noise oscillator which can switch with sine osc
Noise OSC can be played by setting frequency to negative value.

(Therefore a tone with a negative frequency is no longer a rest. Only
a tone with frequency == 0 is a rest.)

Because we need more space due to brown noise table, we use a smaller
table for sine osc.

We also needed some optimization to avoid clicks.